sql fix  psql: id bigint

+++ persistence/src/main/resources/db/migration/V1__schema.sql
@@ -3,14 +3,14 @@ drop table if exists books;
 
 
 create table if not exists books (
-                                     id integer generated always as identity primary key,
+                                     id bigint generated always as identity primary key,
                                      title varchar(255) not null,
     author varchar(255) not null,
     pub_year int not null
     );
 
 create table if not exists comments (
-                                        id integer generated always as identity primary key,
+                                        id bigint generated always as identity primary key,
                                         book_id bigint not null,
                                         author varchar(64) not null,
     text varchar(1000) not null,
